feat(pid,boost): phase boost initiale + PID sur température filtrée
- PHASE_BOOST : pleine puissance (bypass PID) au démarrage du programme,
  jusqu'au décollage thermique (+boostTempRise °C) ou timeout (boostMaxSec s).
  Réduit le dead time initial sans paramétrage complexe.
- FiringProgram : champs enableBoost, boostTempRise (défaut 5°C), boostMaxSec (défaut 60s)
- PID : bascule sur filteredTemp (EMA, double) au lieu de currentTemp brut —
  atténue le bruit de quantification du MAX6675 sur le terme D
- programs.html : section boost dans le formulaire (seuil + durée max)

+++ src/data_logger.cpp
@@ -35,7 +35,7 @@ extern double         currentTemp;
 extern double         targetTemp;
 extern double         rawTemp;
 extern double         pidOutput;
-extern float          filteredTemp;
+extern double         filteredTemp;
 extern float          pidError;
 extern float          pidPterm;
 extern float          pidIterm;
